custom knobs update
Prepared to implement custom knobs and styling to the volume knob and oscillator component. The custom knob code crashes the plugin interface in Reaper with an error code on a call to the getWidth() function. I thing it's a problem with the way the dimensions are defined in Oscillator >> resized(). Default sliders and knobs are being using. Volume knob turns input to zero.